Use Back and Close buttons in Modal
