+++ src/react-components/input/BackButton.scss
@@ -1,0 +1,3 @@
+:local(.back-button) {
+  margin-left: -8px;
+}
