+++ src/react-components/input/CloseButton.js
@@ -1,0 +1,11 @@
+import React from "react";
+import { IconButton } from "../input/IconButton";
+import { ReactComponent as CloseIcon } from "../icons/Close.svg";
+
+export function CloseButton(props) {
+  return (
+    <IconButton {...props}>
+      <CloseIcon width={16} height={16} />
+    </IconButton>
+  );
+}
