+++ src/react-components/modal/Modal.js
@@ -2,6 +2,16 @@ import React from "react";
 import PropTypes from "prop-types";
 import classNames from "classnames";
 import styles from "./Modal.scss";
+import { IconButton } from "../input/IconButton";
+import { ReactComponent as CloseIcon } from "../icons/Close.svg";
+
+export function CloseButton(props) {
+  return (
+    <IconButton {...props} className={styles.sidebarButton}>
+      <CloseIcon width={16} height={16} />
+    </IconButton>
+  );
+}
 
 export function Modal({ title, beforeTitle, afterTitle, children, contentClassName, className, disableFullscreen }) {
   return (
