+++ src/react-components/room/AvatarSettingsSidebar.js
@@ -1,6 +1,8 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { Sidebar, BackButton, CloseButton } from "../sidebar/Sidebar";
+import { Sidebar } from "../sidebar/Sidebar";
+import { CloseButton } from "../input/CloseButton";
+import { BackButton } from "../input/BackButton";
 import { AvatarSettingsContent } from "./AvatarSettingsContent";
 
 export function AvatarSettingsSidebar({ className, showBackButton, onBack, onClose, ...rest }) {
