+++ src/react-components/room/AvatarSetupModal.js
@@ -1,22 +1,12 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { Modal } from "../modal/Modal";
-import { ReactComponent as ChevronBackIcon } from "../icons/ChevronBack.svg";
-import { IconButton } from "../input/IconButton";
+import { BackButton } from "../input/BackButton";
 import { AvatarSettingsContent } from "./AvatarSettingsContent";
 
 export function AvatarSetupModal({ className, onBack, ...rest }) {
   return (
-    <Modal
-      title="Avatar Setup"
-      beforeTitle={
-        <IconButton onClick={onBack}>
-          <ChevronBackIcon />
-          <span>Back</span>
-        </IconButton>
-      }
-      className={className}
-    >
+    <Modal title="Avatar Setup" beforeTitle={<BackButton onClick={onBack} />} className={className}>
       <AvatarSettingsContent {...rest} />
     </Modal>
   );
