+++ src/react-components/room/ObjectsSidebar.js
@@ -2,7 +2,8 @@ import React from "react";
 import PropTypes from "prop-types";
 import classNames from "classnames";
 import styles from "./ObjectsSidebar.scss";
-import { Sidebar, CloseButton } from "../sidebar/Sidebar";
+import { Sidebar } from "../sidebar/Sidebar";
+import { CloseButton } from "../input/CloseButton";
 import { ButtonListItem } from "../layout/List";
 import listStyles from "../layout/List.scss";
 import { ReactComponent as ObjectIcon } from "../icons/Object.svg";
