+++ src/react-components/sidebar/Sidebar.js
@@ -2,46 +2,6 @@ import React from "react";
 import PropTypes from "prop-types";
 import classNames from "classnames";
 import styles from "./Sidebar.scss";
-import { IconButton } from "../input/IconButton";
-import { ReactComponent as ChevronBackIcon } from "../icons/ChevronBack.svg";
-import { ReactComponent as CloseIcon } from "../icons/Close.svg";
-
-export function BackButton({ children, ...rest }) {
-  return (
-    <IconButton {...rest} className={styles.backButton}>
-      <ChevronBackIcon />
-      <span>{children}</span>
-    </IconButton>
-  );
-}
-
-BackButton.propTypes = {
-  children: PropTypes.node
-};
-
-BackButton.defaultProps = {
-  children: "Back"
-};
-
-export function CloseButton(props) {
-  return (
-    <IconButton {...props} className={styles.sidebarButton}>
-      <CloseIcon width={16} height={16} />
-    </IconButton>
-  );
-}
-
-export function SidebarButton({ children, ...rest }) {
-  return (
-    <IconButton {...rest} className={styles.sidebarButton}>
-      {children}
-    </IconButton>
-  );
-}
-
-SidebarButton.propTypes = {
-  children: PropTypes.node
-};
 
 export function Sidebar({ title, beforeTitle, afterTitle, children, contentClassName, className }) {
   return (
